+++ server/src/main/java/org/allaymc/server/command/defaults/StructureCommand.java
@@ -2,7 +2,7 @@ package org.allaymc.server.command.defaults;
 
 import org.allaymc.api.command.SenderType;
 import org.allaymc.api.command.tree.CommandTree;
-import org.allaymc.api.world.Structure;
+import org.allaymc.api.world.StructureFile;
 import org.cloudburstmc.nbt.NbtMap;
 import org.cloudburstmc.nbt.NbtUtils;
 import org.joml.Vector3d;
@@ -56,7 +56,7 @@ public class StructureCommand extends VanillaCommand {
                         return context.fail();
                     }
 
-                    var structure = Structure.pick(
+                    var structure = StructureFile.pick(
                             player.getDimension(),
                             (int) start.x, (int) start.y, (int) start.z,
                             sizeX, sizeY, sizeZ,
@@ -104,7 +104,7 @@ public class StructureCommand extends VanillaCommand {
                         return context.fail();
                     }
 
-                    var structure = Structure.formNBT(nbt);
+                    var structure = StructureFile.formNBT(nbt);
                     structure.place(player.getDimension(), (int) pos.x, (int) pos.y, (int) pos.z);
 
                     context.addOutput("Structure " + filePath + " is placed");